+++ main/src/main/java/net/oneandone/stool/stage/Stage.java
@@ -465,7 +465,7 @@ public class Stage {
         return out;
     }
 
-    public void start(Console console, Map<String, String> environment, Map<String, Integer> selection) throws Exception {
+    public void start(Console console, int http, int https, Map<String, String> environment, Map<String, Integer> selection) throws Exception {
         Engine engine;
         String container;
         Engine.Status status;
@@ -484,7 +484,7 @@ public class Stage {
             for (Map.Entry<FileNode, String> mount : mounts.entrySet()) {
                 console.verbose.println("  " + mount.getKey().getAbsolute() + "\t -> " + mount.getValue());
             }
-            hostPorts = session.pool().allocate(this, image.app);
+            hostPorts = session.pool().allocate(this, image.app, http, https);
             container = engine.containerCreate(image.id,  getName() + "." + session.configuration.hostname,
                     OS.CURRENT == OS.MAC /* TODO: why */, 1024L * 1024 * configuration.memory, null, null,
                     hostPorts.toHostLabels(), environment, mounts, image.ports.map(hostPorts));